requirements.txt
django-admin
requests
pillow
channels
uvicorn
websockets
